+++ src/HttpClient/Plugin/Authentication.php
@@ -1,0 +1,65 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the DigitalOceanV2 library.
+ *
+ * (c) Antoine Corcy <contact@sbin.dk>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace DigitalOceanV2\HttpClient\Plugin;
+
+use Http\Client\Common\Plugin;
+use Http\Promise\Promise;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+
+/**
+ * A plugin to add authentication to the request.
+ *
+ * @internal
+ *
+ * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ * @author Graham Campbell <graham@alt-three.com>
+ */
+final class Authentication implements Plugin
+{
+    /**
+     * The authorization header.
+     *
+     * @var string
+     */
+    private $header;
+
+    /**
+     * Create a new authentication plugin instance.
+     *
+     * @param string $token
+     *
+     * @return void
+     */
+    public function __construct(string $token)
+    {
+        $this->header = \sprintf('Bearer %s', $token);
+    }
+
+    /**
+     * Handle the request and return the response coming from the next callable.
+     *
+     * @param \Psr\Http\Message\RequestInterface                     $request
+     * @param callable(RequestInterface): Promise<ResponseInterface> $next
+     * @param callable(RequestInterface): Promise<ResponseInterface> $first
+     *
+     * @return \Http\Promise\Promise<ResponseInterface>
+     */
+    public function handleRequest(RequestInterface $request, callable $next, callable $first): Promise
+    {
+        $request = $request->withHeader('Authorization', $this->header);
+
+        return $next($request);
+    }
+}