+++ src/HttpClient/Plugin/ExceptionThrower.php
@@ -1,0 +1,79 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the DigitalOceanV2 library.
+ *
+ * (c) Antoine Corcy <contact@sbin.dk>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace DigitalOceanV2\HttpClient\Plugin;
+
+use DigitalOceanV2\Exception\ApiLimitExceededException;
+use DigitalOceanV2\Exception\ErrorException;
+use DigitalOceanV2\Exception\RuntimeException;
+use DigitalOceanV2\Exception\ValidationFailedException;
+use DigitalOceanV2\HttpClient\Message\ResponseMediator;
+use Http\Client\Common\Plugin;
+use Http\Promise\Promise;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+
+/**
+ * A plugin to throw bitbucket exceptions.
+ *
+ * @internal
+ *
+ * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ * @author Fabien Bourigault <bourigaultfabien@gmail.com>
+ * @author Graham Campbell <graham@alt-three.com>
+ */
+final class ExceptionThrower implements Plugin
+{
+    /**
+     * Handle the request and return the response coming from the next callable.
+     *
+     * @param \Psr\Http\Message\RequestInterface                     $request
+     * @param callable(RequestInterface): Promise<ResponseInterface> $next
+     * @param callable(RequestInterface): Promise<ResponseInterface> $first
+     *
+     * @return \Http\Promise\Promise<ResponseInterface>
+     */
+    public function handleRequest(RequestInterface $request, callable $next, callable $first): Promise
+    {
+        return $next($request)->then(function (ResponseInterface $response): ResponseInterface {
+            $status = $response->getStatusCode();
+
+            if ($status >= 400 && $status < 600) {
+                throw self::createException($status, ResponseMediator::getErrorMessage($response) ?? $response->getReasonPhrase());
+            }
+
+            return $response;
+        });
+    }
+
+    /**
+     * Create an exception from a status code and error message.
+     *
+     * @param int    $status
+     * @param string $message
+     *
+     * @return ErrorException|RuntimeException
+     */
+    private static function createException(int $status, string $message)
+    {
+        if (400 === $status || 422 === $status) {
+            return new ValidationFailedException($message, $status);
+        }
+
+        if (429 === $status) {
+            return new ApiLimitExceededException($message, $status);
+        }
+
+        return new RuntimeException($message, $status);
+    }
+}