+++ src/HttpClient/Plugin/History.php
@@ -1,0 +1,72 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the DigitalOceanV2 library.
+ *
+ * (c) Antoine Corcy <contact@sbin.dk>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace DigitalOceanV2\HttpClient\Plugin;
+
+use Http\Client\Common\Plugin\Journal;
+use Psr\Http\Client\ClientExceptionInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+
+/**
+ * A plugin to remember the last response.
+ *
+ * @internal
+ *
+ * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ * @author Graham Campbell <graham@alt-three.com>
+ */
+final class History implements Journal
+{
+    /**
+     * The last response.
+     *
+     * @var \Psr\Http\Message\ResponseInterface|null
+     */
+    private $lastResponse;
+
+    /**
+     * Get the last response.
+     *
+     * @return \Psr\Http\Message\ResponseInterface|null
+     */
+    public function getLastResponse()
+    {
+        return $this->lastResponse;
+    }
+
+    /**
+     * Record a successful call.
+     *
+     * @param \Psr\Http\Message\RequestInterface  $request
+     * @param \Psr\Http\Message\ResponseInterface $response
+     *
+     * @return void
+     */
+    public function addSuccess(RequestInterface $request, ResponseInterface $response): void
+    {
+        $this->lastResponse = $response;
+    }
+
+    /**
+     * Record a failed call.
+     *
+     * @param \Psr\Http\Message\RequestInterface        $request
+     * @param \Psr\Http\Client\ClientExceptionInterface $exception
+     *
+     * @return void
+     */
+    public function addFailure(RequestInterface $request, ClientExceptionInterface $exception): void
+    {
+    }
+}