+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGE LOG
 * Added void return types to void methods
 * Dropped support for PHP 7.1
 * Switched to PSR-17 and PSR-18
+* Made builder class final
 
 
 ## 3.2.1 (27/11/2020)